rtc: enable irq on level instead of edge
RTC ISR is not getting called in deepsleep even interrupt is fired.
As the interrupt is configured as a edge, only a transition will be
triggering an interrupt from the APIC point of view.
When the RTC wakes up the SOC, the transition is occuring prior to the
activation / initialization of the APIC, hence the need to configure in
level instead of edge

Change-Id: If5273bf9884273f14eef8570ce50b4b38400a851
Tracked-On: https://jira.ndg.intel.com/browse/FIRE-6398
Signed-off-by: Tadikonda Venkateswara Rao <venkateswara.raox.tadikonda@intel.com>
Reviewed-on: https://android.intel.com/523483
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: D'alton, Alexandre <alexandre.dalton@intel.com>
Reviewed-by: Chevalier, Richard <richard.chevalier@intel.com>