led: handle semaphore give on multiple led play request
As we were continously waiting on semaphore when led is played
multiple times(PATTERN_OVERRIDE), queue was getting full with
multiple request which was leading to panic.

Release the semaphore in pattern override case as
ui_play_led_callback will not be called, in order to resolve
the above issue.

Change-Id: I1d77cefe77591817784aa2de7e78d1486e64dc9e
Tracked-On: https://jira.ndg.intel.com/browse/FIRE-7692
Signed-off-by: Mourya Abhishek <abhishekx.mourya@intel.com>
Reviewed-on: https://android.intel.com/543170
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Cheverneuil, FredericX <fredericx.cheverneuil@intel.com>
Reviewed-by: Chereau, Fabien <fabien.chereau@intel.com>
Reviewed-on: https://android.intel.com/543444
Reviewed-by: Du, Alek <alek.du@intel.com>